pyproject.toml
[project]
name = "dspy-tryout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "datasets>=4.0.0",
    "dspy @ git+https://github.com/stanfordnlp/dspy.git@c1c27b4",
    "langfuse>=3.5.0",
    "litellm==1.75.9",
    "openinference-instrumentation-dspy>=0.1.29",
    "python-dotenv>=1.1.1",
]
